Use smallint for sort column on stories table

+++ app/database/migrations/2013_07_24_204822_add_sort_column_to_stories_table.php
@@ -13,7 +13,7 @@ class AddSortColumnToStoriesTable extends Migration {
     public function up()
     {
         Schema::table('stories', function(Blueprint $table) {
-            $table->integer('sort')->unsigned()->default(1);
+            $table->smallInteger('sort')->unsigned()->default(1);
             $table->unique(['sort', 'issue_id']);
         });
     }
